Refactor dependency insight renderer for readability
This commit largely refactors the dependency insight report
infrastructure to make it more understandable. The same should
probably be done for the ascii reporter, but it's not as widely
used. This is the reason why the dependency graphs builder has
a `complete` method, instead of writing the legend directly
when all nodes are rendered: it is called multiple times,
instead of once.